DsATAC.cellranger

Create a DsATAC dataset from multiple input files output by 10x cellranger

DsATAC.cellranger(
  sampleAnnot,
  sampleDirPrefixCol,
  genome,
  dataDir = "",
  regionSets = NULL,
  addPeakRegions = TRUE,
  sampleIdCol = sampleDirPrefixCol,
  keepInsertionInfo = FALSE,
  diskDump.fragments = keepInsertionInfo
)

Arguments

sampleAnnot

data.frame specifying the sample annotation table

sampleDirPrefixCol

column name specifying the directory prefix for each sample in the sample annotation table

genome

genome assembly

dataDir

directory where the files are located

regionSets

a list of GRanges objects which contain region sets over which count data will be aggregated

addPeakRegions

should a merged set of peaks be created as one of the region sets (merged, non-overlapping peaks of width=500bp from the peaks of individual samples)

sampleIdCol

column name or index in the sample annotation table containing unique sample identifiers

keepInsertionInfo

flag indicating whether to maintain the insertion information in the resulting object.

diskDump.fragments

Keep fragment coordinates stored on disk rather than in main memory. This saves memory, but increases runtime and I/O.

Value

DsATAC object

Author

Fabian Mueller