DsATAC.snakeATAC

Create a DsATAC dataset from multiple input files output by snakeATAC

DsATAC.snakeATAC(sampleAnnot, filePrefixCol, genome, dataDir = "",
  regionSets = NULL, sampleIdCol = filePrefixCol, type = "insBam",
  diskDump = FALSE, keepInsertionInfo = TRUE, bySample = FALSE,
  pairedEnd = TRUE)

Arguments

sampleAnnot
data.frame specifying the sample annotation table

filePrefixCol
column name specifying the file prefix for each sample in the sample annotation table. If dataDir is not empty (i.e. not "") filenames are assumed to be relative to that directory and a corresponding filename suffix will be appended

genome
genome assembly

dataDir
directory where the files are located. If it is the empty character (""; default) it is assumed that filePrefixCol specifies the full path to the input files

regionSets
a list of GRanges objects which contain region sets over which count data will be aggregated

sampleIdCol
column name or index in the sample annotation table containing unique sample identifiers

type
input data type. Currently only "insBed" (insertion beds), "insBed" (insertion info inferred from bam files (aligned reads); default) and "bam" (aligned reads) are supported

diskDump
should large data objects (count matrices, fragment data, ...) be disk-backed to save main memory

keepInsertionInfo
flag indicating whether to maintain the insertion information in the resulting object. Only relevant when type=="insBam".

bySample
process sample-by-sample to save memory (currently only has an effect for type=="insBam")

pairedEnd
is the input data paired-end? Only relevant when type=="insBam".

Value

DsATAC object