callPeaks-methods

Performs peak calling based on insertion sites

# S4 method for DsATAC
callPeaks(
  .object,
  samples = getSamples(.object),
  method = "macs2_summit_fw_no",
  methodOpts = list(macs2.exec = "macs2", macs2.params = c("--shift", "-75",
    "--extsize", "150", "-p", "0.01"), fixedWidth = 250)
)

Arguments

.object
DsATAC object

samples
sample identifiers for which peak calling is performed

method
peak calling method. Currently only 'macs2_summit_fw_no' is supported. See details section.

methodOpts
list of other options depending on the 'method' parameter (see details section).

Value

GRangesList of peak coordinates for each sample

Details

The following methods are currently supported

'macs2_summit_fw_no'

Fixed-width, non-overlapping peaks based on MACS2 summit calls: 1. Call peaks using system call to MACS2. You can specify the MACS2 executable in methodOpts$macs2.exec. 2. Identify peak summits 3. extend peak summits on each side by a number of basepairs (specified in methodOpts$fixedWidth; default: 250bp) to obtain unified peak widths 4. Find non-overlapping peaks by taking the peak with the best MACS2 score from each set of partially overlapping peaks

Author

Fabian Mueller