computeDiffAcc.rnb.nome

computes differential accessibility for NOMe datasets using RnBeads functionality

computeDiffAcc.rnb.nome(dsn, cmpCols, regionTypes = getRegionTypes(dsn),
  covgThres = 5L, allPairs = TRUE, adjPairCols = NULL,
  adjCols = NULL, skipSites = FALSE,
  disk.dump = rnb.getOption("disk.dump.big.matrices"),
  disk.dump.dir = tempfile(pattern = "diffMethTables_"), ...)

Arguments

dsn
DsNOMe object

cmpCols
column names of the sample annotation of the dataset that will be used for comparison

regionTypes
which region types should be processed for differential analysis.

covgThres
coverage threshold for computing the summary statistics. See RnBeads::computeDiffTab.extended.site for details.

allPairs
Logical indicating whether all pairwise comparisons should be conducted, when more than 2 groups are present

adjPairCols
argument passed on to rnb.sample.groups. See its documentation for details.

adjCols
not used yet

skipSites
flag indicating whether differential methylation in regions should be computed directly and not from sites. This leads to skipping of site-specific differential methylation

disk.dump
Flag indicating whether the resulting differential methylation object should be file backed, ie.e the matrices dumped to disk

disk.dump.dir
disk location for file backing of the resulting differential methylation object. Only meaningful if disk.dump=TRUE. must be a character specifying an NON-EXISTING valid directory.

...
arguments passed on to binary differential methylation calling. See RnBeads::computeDiffTab.extended.site for details.

Value

an RnBDiffMeth object. See class description for details.