dimRed_UMAP-methods

Retrieve dimension reduction embedding and object using UMAP

# S4 method for DsATACsc
dimRed_UMAP(
  .object,
  regions,
  tfidf = TRUE,
  pcs = 1:50,
  normPcs = FALSE,
  umapParams = list(distMethod = "euclidean", min_dist = 0.5, n_neighbors = 25),
  rmDepthCor = 1
)

Arguments

.object
DsATACsc object

regions
character string specifying the region type to retrieve the UMAP coordinates from. Alternatively, a GRanges object specifying coordinates that fragment counts will be aggregated over

tfidf
normalize the counts using TF-IDF transformation

pcs
components to use to compute the SVD

normPcs
flag indicating whether to apply z-score normalization to PCs for each cell

umapParams
parameters to compute UMAP coordinates (passed on to muRtools::getDimRedCoords.umap and further to uwot::umap)

rmDepthCor
correlation cutoff to be used to discard principal components associated with fragment depth (all iterationa). By default (value >=1) no filtering will be applied.

Value

an S3 object containing dimensionality reduction results

Details

The output object includes the final singular values/principal components (result$pcaCoord), the low-dimensional coordinates (result$umapCoord) as well as region set that provided the basis for the dimension reduction (result$regionGr).

Author

Fabian Mueller