getComparisonInfo

retrieve the comparison information for an DsAcc object. Analogous to RnBeads::get.comparison.info

getComparisonInfo(
  dsa,
  cmpNames = NULL,
  regionTypes = getRegionTypes(dsa),
  allPairs = TRUE,
  adjPairCols = NULL,
  minGrpSize = 1L,
  maxGrpCount = NULL
)

Arguments

dsa
DsAcc object

cmpNames
column names of the sample annotation of the dataset that will be used for comparison

regionTypes
which region types should be processed for differential analysis.

allPairs
Logical indicating whether all pairwise comparisons should be conducted, when more than 2 groups are present

adjPairCols
argument passed on to rnb.sample.groups. See its documentation for details.

minGrpSize
Minimum number of samples required to form a group in comparison

maxGrpCount
maximum number of groups to consider for comparisons

Value

a list containing one element for each comparison to be conducted. Each element is again a list containing:

comparison

the name of the comparison

pheno.colname

the column name of the sample annotation table the comparison is derived from

group.names

the names of the two groups being compared

group.inds

the sample indices of the samples belonging to the two groups

paired

flag indicating whether paired analysis is conducted

adj.sva

flag indicating whether adjustment for SVA is conducted

adj.celltype

flag indicating whether adjustment for cell type is conducted

adjustment.table

the covariate adjustment table. NULL if the comparison is not adjusted

region.types

the region types applicable to the analysis

Author

Fabian Mueller