getDiffAcc-methods

Compute differential accessibility

# S4 method for DsATAC
getDiffAcc(
  .object,
  regionType,
  comparisonCol,
  grp1Name = NULL,
  grp2Name = NULL,
  adjustCols = character(0),
  method = "DESeq2",
  diffObj = NULL
)

Arguments

.object

DsATAC object

regionType

character string specifying the region type

comparisonCol

column name in the sample annotation table to base the comparison on

grp1Name

name of the first group in the comparison. if not specified, it will be taken as the first factor level specified in the sample annotation table in 'comparisonCol'.

grp2Name

name of the second group (reference) in the comparison. if not specified, it will be taken as the first factor level specified in the sample annotation table in 'comparisonCol'.

adjustCols

column names in the sample annotation potentially used to create the design matrix

method

Method for determining differential accessibility. Currently only 'DESeq2' is supported

diffObj

optional differential object to avoid computing it for each comparison and thus reduce runtime

Value

a data.frame containing differential accessibility information

Author

Fabian Mueller

Examples

if (FALSE) {
dsa <- ChrAccRex::loadExample("dsAtac_ia_example")
daTab <- getDiffAcc(dsa, "IA_prog_peaks", "stimulus", grp1Name="S", grp2Name="U", adjustCols=c("cellType", "donor"))
}