getMotifClustering

Retrieve motif clustering of TF motifs

getMotifClustering(k = 0, distM = NULL, assembly = "hg38",
  motifs = "jaspar", clusterMethod = "pam")

Arguments

k
number of clusters. k<1 will result in an automatically selected clustering which is precomputed and stored in ChrAccR. For distMethod=="jaspar" and clusterMethod=="pam" this corresponds to the k corresponding to the best silhouette value before a drop (in the silhouette elbow-curve) occurs

distM
distance matrix (dist object) containing motif dissimilarities/distances. Only required if k>0.

assembly
genome assembly for which the motifs dissimilarity should be retrieved. Only the species information of the assembly is really relevant. Can be "vert" for all vertebrate motifs. Only required if for automatic mode (i.e. k<1).

motifs
either a character string (currently only "jaspar" is supported) or an object containing PWMs that can be used by motifmatchr::matchMotifs (PWMatrixList object)

clusterMethod
method to be used for motif clustering (currently only 'pam' (PAM - partitioning around medoids) is supported)

distMethod
method of dissimilarity quantification. Currently only 'jaspar' (retrieve motif similarities from the annotation of the JASPAR website) is supported.

Value

a list structure containing the clustering result