getTssEnrichment-methods

Get TSS enrichment data and plot

# S4 method for DsATAC
getTssEnrichment(
  .object,
  sampleId,
  tssGr = NULL,
  flank = 2000L,
  normTailW = 100L,
  smoothW = 25L,
  silent = FALSE
)

Arguments

.object
DsATAC object

sampleId
sample to be plotted

tssGr
GRanges object containing TSS coordinates or NULL to get default set from annotation package

flank
number of bases flanking each TSS that will be added on each side

normTailW
number of bases on each side whose counts will be used to normalize the data

smoothW
radius of the window (in bp) that will be used to smooth the data, i.e. the total width of the smoothing window will be twice that number

silent
limit log messages

Value

a list containing TSS enrichment data and a ggplot object containing TSS enrichment plot

Author

Fabian Mueller