hmSeqLogo

Draw a sequence motif logo in a Complex Heatmap using grid. adapted from seqLogo::seqLogo()

hmSeqLogo(pwm, x, y, width, height, ic.scale = TRUE)

Arguments

pwm
PWM (from TFBSTools package)

x
x center coordinate where the motif should be drawn

width
drawing width

height
drawing height

ic.scale
logical If TRUE, the height of each column is proportional to its information content. Otherwise, all columns have the same height.

x
y center coordinate where the motif should be drawn

Value

Draws the motif