setConfigElement

Set a configuration item to a given value

setConfigElement(name, value)

Arguments

name
name of the config item

value
value of the config item

Value

nothing of particular interest.

Options used by the package

tmpDir = temdir()

Directory for temporary files. Must be existing.

cleanMem = TRUE

During runtime, regularly clean-out the memory in order to reduce memory overuse

colorSchemes

named list of DISCRETE color schemes to be used for plotting. Each element should be a named vector specifying colors for groups/annotations.

colorSchemesCont

named list of CONTINOUS color schemes to be used for plotting. Each element should be a vector specifying a range of colors.

geneModelVersions

Gene model versions to be used for various genomes

analysisName = "ChrAccR analysis"

A title for the analysis (a string).

regionTypes

Region types to be used in the analysis

chromVarRegionTypes = NULL

Region types to be used for chromVar analysis. If NULL (default), ChrAccR will automatically look for region types with the keyword "peak" in their name.

chromVarMotifs = "jaspar_vert"

Character vector of names of TF motif sets to be used in ChromVAR analyses. By default the vertebrate set of the JASPAR database will be used.

chromVarMotifNamesForDimRed

Names of motifs to be used for dimension reduction plots in the reports. [only relevant for single-cell data]

genesOfInterest

Names of genes of interest to be highlighted in the reports (e.g. dimension reduction) in the reports. [currently only relevant for single-cell data and only when scGeneActivity is activated]

annotationColumns

Sample annotation columns to be used for reporting

annotationMinGroupSize = 2

Minimum size of a group to be used in the reports. Influences which annotation columns are automatically selected for reporting.

annotationMaxGroupCount = NULL

Maximum number of groups to be used in the reports. Influences which annotation columns are automatically selected for reporting. If NULL (default) it will effectively be the number of samples - 1.

doPeakCalling = FALSE

Perform per-sample peak calling and retrieve consensus peak set. Requires that macs2 is installed and can be called from the command line. [for bulk data analysis only]

annotationPeakGroupColumn

Annotation column to base the consensus peak set replication filtering on.

annotationPeakGroupAgreePerc = 1.0

Percent of samples that have to agree to identify consensus peaks. See getConsensusPeakSet for details.

filteringCovgCount = 1L

Minimum insertion count to filter count matrices by. See filterLowCovg,DsATAC-method for details. [for bulk data analysis only]

filteringCovgReqSamples = 0.75

Minimum required samples to apply low coverage filtering to. See filterLowCovg,DsATAC-method for details. [for bulk data analysis only]

filteringSexChroms = FALSE

Flag indicating whether to remove sex chromosomes.

filteringScMinFragmentsPerCell = 1000L

Minimum number of fragments per cell to retain a cell in the analysis. [for single-cell data analysis only]

filteringScMaxFragmentsPerCell = Inf

Maximum number of fragments allowed per cell to retain a cell in the analysis. [for single-cell data analysis only]

filteringScMinTssEnrichment = 6

Minimum TSS enrichment score per cell to retain a cell in the analysis. [for single-cell data analysis only]

normalizationMethod = "quantile"

Normalization method to use for count normalization. Allowed methods include the ones listed in transformCounts,DsATAC-method. [for bulk data analysis only]

exploratoryLogNormCounts = TRUE

Should a log-normalization be applied in the exploratory plot sections of the reports (dimension reduction, heatmaps)

exploratoryNSubsample = 2e6

Number of regions to subsample in exploratory analysis in order to increase computational performance.

differentialColumns

Sample annotation columns to be used for differential testing and reporting

differentialColumns1vsAll

Sample annotation columns to be used for differential testing and reporting in a 1-vs-all group setting. Should be a subset of differentialColumns.

differentialCompNames

Comparison names from which comparison information is derived. Must be in the format of "$GRP1_NAME vs $GRP2_NAME [$ANNOTATION_COLUMN]".

differentialAdjColumns

Sample annotation columns to be adjusted for in differential testing

lolaDbPaths

Precomputed LOLA databases to be used for enrichment analysis. If NULL (default), ChrAccR will download an apropriate core database.

scIterativeLsiRegType

For single-cell analysis only: region type to be used for clustering and dimension reduction using iterative LSI. By default (NULL), ChrAccR will look for a region type named "tiling".

scIterativeLsiParams

Parameters to use for iterative LSI. See iterativeLSI,DsATACsc-method for details.

scGeneActivity = FALSE

For single-cell analysis only: Compute gene activity from accessibility. Possible options are "RBF" for radial-basis-function-weighted count aggregation (default when set to TRUE) or "Cicero" for Cicero correlation-based aggregation

Author

Fabian Mueller